requirements.txt
matplotlib>=3.7.2
numpy>=1.24.2
torchvision>=0.9.0
scikit-learn>=1.3.0
transformers>=4.28.1
pillow>=9.5.0
pandas>=1.3.0
jieba>=0.42.1